+++ love/src/components/NightReport/CreateNightReport.container.jsx
@@ -1,0 +1,65 @@
+/** 
+This file is part of LOVE-frontend.
+
+Developed for the Vera C. Rubin Observatory Telescope and Site Systems.
+
+This program is free software: you can redistribute it and/or modify it under 
+the terms of the GNU General Public License as published by the Free Software 
+Foundation, either version 3 of the License, or at your option) any later version.
+
+This program is distributed in the hope that it will be useful,but WITHOUT ANY
+ WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR 
+ A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License along with 
+this program. If not, see <http://www.gnu.org/licenses/>.
+*/
+
+import React from 'react';
+import { connect } from 'react-redux';
+import SubscriptionTableContainer from 'components/GeneralPurpose/SubscriptionTable/SubscriptionTable.container';
+import { addGroup, removeGroup } from 'redux/actions/ws';
+import CreateNightReport from './CreateNightReport';
+
+export const schema = {
+  description: 'Night report writting service',
+  defaultSize: [77, 32],
+  props: {
+    title: {
+      type: 'string',
+      description: 'Name displayed in the title bar (if visible)',
+      isPrivate: false,
+      default: 'Log Service',
+    },
+    hasRawMode: {
+      type: 'boolean',
+      description: 'Whether the component has a raw mode version',
+      isPrivate: true,
+      default: false,
+    },
+  },
+};
+
+const CreateNightReportContainer = ({ subscribeToStreams, unsubscribeToStreams, ...props }) => {
+  if (props.isRaw) {
+    return <SubscriptionTableContainer subscriptions={props.subscriptions}></SubscriptionTableContainer>;
+  }
+  return (
+    <CreateNightReport subscribeToStreams={subscribeToStreams} unsubscribeToStreams={unsubscribeToStreams} {...props} />
+  );
+};
+
+const mapDispatchToProps = (dispatch) => {
+  const subscriptions = [];
+  return {
+    subscriptions,
+    subscribeToStreams: () => {
+      subscriptions.forEach((s) => dispatch(addGroup(s)));
+    },
+    unsubscribeToStreams: () => {
+      subscriptions.forEach((s) => dispatch(removeGroup(s)));
+    },
+  };
+};
+
+export default connect(null, mapDispatchToProps)(CreateNightReportContainer);