+++ CHANGELOG.rst
@@ -2,6 +2,11 @@
 Version History
 ===============
 
+v5.30.0
+-------
+
+* Add Night Report implementation `<https://github.com/lsst-ts/LOVE-frontend/pull/619>`_
+
 v5.29.3
 -------